+++ lib/oban/crontab/scheduler.ex
@@ -27,7 +27,8 @@ defmodule Oban.Crontab.Scheduler do
       :poll_ref,
       circuit: :enabled,
       reset_timer: nil,
-      poll_interval: :timer.seconds(60)
+      poll_interval: :timer.seconds(60),
+      first_run?: true
     ]
   end
 
@@ -66,9 +67,9 @@ defmodule Oban.Crontab.Scheduler do
     state =
       state
       |> send_poll_after()
-      |> lock_and_enqueue()
+      |> lock_and_insert()
 
-    {:noreply, state}
+    {:noreply, %{state | first_run?: false}}
   end
 
   def handle_info({:EXIT, _pid, error}, %State{} = state) do
@@ -89,13 +90,13 @@ defmodule Oban.Crontab.Scheduler do
     %{state | poll_ref: ref}
   end
 
-  defp lock_and_enqueue(%State{circuit: :disabled} = state), do: state
+  defp lock_and_insert(%State{circuit: :disabled} = state), do: state
 
-  defp lock_and_enqueue(%State{conf: conf, poll_interval: timeout} = state) do
+  defp lock_and_insert(%State{conf: conf, poll_interval: timeout} = state) do
     %Config{repo: repo, log: log} = conf
 
     repo.transaction(
-      fn -> if Query.acquire_lock?(conf, @lock_key), do: enqueue_jobs(conf) end,
+      fn -> if Query.acquire_lock?(conf, @lock_key), do: insert_jobs(state) end,
       log: log,
       timeout: timeout
     )
@@ -105,16 +106,32 @@ defmodule Oban.Crontab.Scheduler do
     exception in trip_errors() -> trip_circuit(exception, __STACKTRACE__, state)
   end
 
-  defp enqueue_jobs(%Config{crontab: crontab, timezone: timezone} = conf) do
+  defp insert_jobs(%State{conf: conf, first_run?: first_run?}) do
+    if first_run?, do: insert_reboot_jobs(conf)
+
+    insert_scheduled_jobs(conf)
+  end
+
+  defp insert_scheduled_jobs(%Config{crontab: crontab, timezone: timezone} = conf) do
     {:ok, datetime} = DateTime.now(timezone)
 
     for {cron, worker, opts} <- crontab, Cron.now?(cron, datetime) do
-      {args, opts} = Keyword.pop(opts, :args, %{})
-
-      opts = unique_opts(worker.__opts__(), opts)
-
-      {:ok, _job} = Query.fetch_or_insert_job(conf, worker.new(args, opts))
+      insert_job(worker, opts, conf)
     end
+  end
+
+  defp insert_reboot_jobs(%Config{crontab: crontab} = conf) do
+    for {cron, worker, opts} <- crontab, Cron.reboot?(cron) do
+      insert_job(worker, opts, conf)
+    end
+  end
+
+  defp insert_job(worker, opts, conf) do
+    {args, opts} = Keyword.pop(opts, :args, %{})
+
+    opts = unique_opts(worker.__opts__(), opts)
+
+    {:ok, _job} = Query.fetch_or_insert_job(conf, worker.new(args, opts))
   end
 
   # Ensure that `unique` is deep merged and the default period has the lowest priority.